third_party/WebKit/LayoutTests/http/tests/security/cookies/websocket/third-party-cookie-blocked.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/websocket/resources/get-request-header.js"></script>
<script src="/websocket/resources/cookie-test-util.js"></script>
<script>
if (window.testRunner) {
  // Block third-party cookies.
  testRunner.setBlockThirdPartyCookies(true);
}

const origin = WS_REMOTE_ORIGIN;

promise_test(t => {
  return setCookie(origin).then(
    () => {
      return connectAndGetRequestHeadersFrom(origin);
    }
  ).then(
    headers => {
      const cookie = headers['cookie'];
      assert_equals(cookie, undefined,
                    'headers should not contain any cookie header.');
      return clear(origin);
    }
  );
}, "Test third-party cookie blocking for WebSockets");
</script>
